+++ tests/testthat/test-step10-export-data.R
@@ -13,6 +13,7 @@ test_that("{shinytest2} recording: checking-module10", {
   app$click("data1-load_prj")
 
   app$set_inputs(modules = "step-10")
+  app$set_inputs(`data10-update_parms-y_var` = "fDOM_QSU")
   app$wait_for_idle()
 
   #check initial plot